Update main.py
Added the start of an inventory system, currenting gold and potions.

Changed the randomness of enemy spawns

Removed faulty quite options.

Added gold and potions to appear under status.

Changed the randomness of oversleeping.

Changed the randomness of leveling up.

Changed the randomness of finding gold coins.

Changed the amount of max hp gained when leveling up to +3