Update main.py
Added color to certain text output.

Worked on the save system(still unfinished)

Changed some typos through out the code.

Worked on printing out a status report when character levels up(print shows "None" for some reason)

Worked on the system to restart the game when character dies, takes you to the mains creen but when selecting the option to start it exits out the program.

Added levelup def, will remove because it does not work.